Dodatocne som ozivil obrazovku GroupTrainingLobby ktora uz obsahuje funkcnu casomieru a funkcne tladicla (pristupne len trenerovi) ktore ju dokazu ovladat, taktiez pribudol list aktualne prihlasenych pouzivatelov v treningu (v realnom case) a stopky sa zatial len pre testovacie ucely vedia spustat vsetkym naraz (mozno so milisekundovym oneskorenim) ale neskor sa budu ucastnici treningu presuvat na vlastne obrazovky s ich statistikami

+++ app/src/main/java/com/example/smartfit/screens/GroupTrainingLobby.kt
@@ -3,14 +3,16 @@ package com.example.smartfit.screens
 import android.graphics.Bitmap
 import android.graphics.Color
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Delete
@@ -18,6 +20,7 @@ import androidx.compose.material.icons.filled.QrCode
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
@@ -32,26 +35,51 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import com.example.smartfit.components.CustomButton
+import com.example.smartfit.components.CustomProfilePictureFrame
 import com.example.smartfit.components.Heading1
+import com.example.smartfit.components.Heading2
+import com.example.smartfit.components.NormalText
+import com.example.smartfit.components.StopWatch
+import com.example.smartfit.data.GroupTraining
+import com.example.smartfit.data.User
+import com.example.smartfit.data.frameColors
+import com.example.smartfit.data.trainingList
 import com.google.zxing.BarcodeFormat
 import com.google.zxing.qrcode.QRCodeWriter
+import kotlinx.coroutines.delay
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun GroupTrainingLobby(
-    groupTrainingId: String,
-    onDeleteClick: () -> Unit
+    chosenGroupTraining: GroupTraining,
+    currentUser: User,
+    allTrainingParticipants: List<User>,
+    onCheckAllTrainingInfo: () -> Boolean,
+    trainingState: (Int) -> Unit, //0 - initial, 1 - start, 2 - pause, 3 - end
+    onDeleteClick: () -> Unit,
+    onEndGroupTrainingClick: (GroupTraining) -> Unit,
 ) {
+
+    var groupTraining by remember { mutableStateOf(chosenGroupTraining) }
+
 
     val context = LocalContext.current
     var qrBitmap by remember { mutableStateOf<Bitmap?>(null) }
 
-    LaunchedEffect(groupTrainingId) {
+    val fullscreenQrState = remember { mutableStateOf(false) }
+
+    val stopWatch = remember { StopWatch() }
+    val isStopWatchRunning = remember { mutableStateOf(stopWatch.isRunning()) }
+
+
+    LaunchedEffect(chosenGroupTraining.id) {
         val writer = QRCodeWriter()
-        val bitMatrix = writer.encode(groupTrainingId, BarcodeFormat.QR_CODE, 512, 512)
+        val bitMatrix = writer.encode(chosenGroupTraining.id, BarcodeFormat.QR_CODE, 512, 512)
         val width = bitMatrix.width
         val height = bitMatrix.height
         val bitmap = Bitmap.createBitmap(width, height, Bitmap.Config.RGB_565)
@@ -61,11 +89,45 @@ fun GroupTrainingLobby(
             }
         }
         qrBitmap = bitmap
+
+        //trainingState(0)
+
+        if (currentUser.id == chosenGroupTraining.trainerId) {
+            fullscreenQrState.value = true
+        }
+        while (true) {
+            if (onCheckAllTrainingInfo()) {
+                delay(10)
+            } else {
+                break
+            }
+        }
+    }
+
+    when (chosenGroupTraining.trainingState) {
+        0 -> {
+            stopWatch.reset()
+            isStopWatchRunning.value = stopWatch.isRunning()
+        }
+
+        1 -> {
+            stopWatch.start()
+            isStopWatchRunning.value = stopWatch.isRunning()
+        }
+
+        2 -> {
+            stopWatch.pause()
+            isStopWatchRunning.value = stopWatch.isRunning()
+        }
+
+        3 -> {
+            stopWatch.pause()
+            groupTraining = groupTraining.copy(trainingDuration = stopWatch.getTimeMillis())
+            onEndGroupTrainingClick(groupTraining)
+        }
     }
 
     qrBitmap?.let { bitmap ->
-
-        val fullscreenQrState = remember { mutableStateOf(true) }
 
         Scaffold(
             topBar = {
@@ -84,16 +146,84 @@ fun GroupTrainingLobby(
                                 }
                             }
 
-                            IconButton(onClick = { onDeleteClick() }) {
-                                Icon(
-                                    imageVector = Icons.Filled.Delete,
-                                    contentDescription = "Delete training",
-                                    tint = MaterialTheme.colorScheme.error
-                                )
+                            if (currentUser.id == chosenGroupTraining.trainerId) {
+                                IconButton(onClick = { onDeleteClick() }) {
+                                    Icon(
+                                        imageVector = Icons.Filled.Delete,
+                                        contentDescription = "Delete training",
+                                        tint = MaterialTheme.colorScheme.error
+                                    )
+                                }
                             }
                         }
                     }
                 )
+            },
+            bottomBar = {
+
+                if (currentUser.id == chosenGroupTraining.trainerId) {
+                    Row() {
+                        if (stopWatch.getTimeMillis() == 0L && !isStopWatchRunning.value) {
+                            CustomButton(
+                                modifier = Modifier
+                                    .padding(20.dp)
+                                    .weight(1f),
+                                onClick = {
+//                                    stopWatch.start()
+//                                    isStopWatchRunning.value = stopWatch.isRunning()
+                                    trainingState(1)
+                                },
+                                containerColor = MaterialTheme.colorScheme.primary,
+                                buttonText = "Sputstit"
+                            )
+                        } else {
+
+                            if (isStopWatchRunning.value) {
+                                CustomButton(
+                                    modifier = Modifier
+                                        .padding(20.dp)
+                                        .weight(1f),
+                                    onClick = {
+//                                        stopWatch.pause()
+//                                        isStopWatchRunning.value = stopWatch.isRunning()
+                                        trainingState(2)
+                                    },
+                                    containerColor = MaterialTheme.colorScheme.secondary,
+                                    buttonText = "Pozastavit"
+                                )
+                            } else {
+                                CustomButton(
+                                    modifier = Modifier
+                                        .padding(20.dp)
+                                        .weight(1f),
+                                    onClick = {
+//                                        stopWatch.start()
+//                                        isStopWatchRunning.value = stopWatch.isRunning()
+                                        trainingState(1)
+                                    },
+                                    containerColor = MaterialTheme.colorScheme.tertiary,
+                                    buttonText = "Sputstit"
+                                )
+                            }
+
+                            CustomButton(
+                                modifier = Modifier
+                                    .padding(20.dp)
+                                    .weight(1f),
+                                onClick = {
+                                    //stopWatch.pause()
+                                    trainingState(3)
+//                                    groupTraining =
+//                                        groupTraining.copy(trainingDuration = stopWatch.getTimeMillis())
+                                    //onEndGroupTrainingClick(groupTraining)
+
+                                },
+                                containerColor = MaterialTheme.colorScheme.error,
+                                buttonText = "Dokoncit"
+                            )
+                        }
+                    }
+                }
             }
 
         ) { innerPadding ->
@@ -106,7 +236,8 @@ fun GroupTrainingLobby(
                     Box(
                         modifier = Modifier
                             .fillMaxSize()
-                            .background(androidx.compose.ui.graphics.Color.Black.copy(alpha = 0.7f)),
+                        //.background(androidx.compose.ui.graphics.Color.Black.copy(alpha = 0.7f))
+                        ,
                         contentAlignment = Alignment.BottomEnd
                     ) {
                         Column(Modifier.fillMaxWidth()) {
@@ -120,7 +251,7 @@ fun GroupTrainingLobby(
                                     Icon(
                                         Icons.Default.Close,
                                         "Close qr code fullscreen icon",
-                                        tint = MaterialTheme.colorScheme.surface
+                                        tint = MaterialTheme.colorScheme.onSurface
                                     )
                                 }
                             }
@@ -137,11 +268,46 @@ fun GroupTrainingLobby(
                 Box(
                     modifier = Modifier
                         .fillMaxSize()
-                        .padding(horizontal = 20.dp),
+                        .padding(horizontal = 15.dp),
                     contentAlignment = Alignment.TopCenter
                 ) {
 
+                    Column {
 
+                        Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Start) {
+                            if (chosenGroupTraining.name.isNotEmpty()) {
+                                Heading2(
+                                    "(${chosenGroupTraining.name})",
+                                    fontWeight = FontWeight.Medium
+                                )
+                                Spacer(Modifier.padding(3.dp))
+                            }
+                            Heading2(trainingList[chosenGroupTraining.trainingIndex].name)
+                            Spacer(Modifier.padding(3.dp))
+                            Heading2(stopWatch.getCustomFormattedTime())
+                        }
+
+                        LazyColumn {
+
+                            items(allTrainingParticipants) { participant ->
+
+                                ListItem(
+                                    headlineContent = { NormalText(participant.displayName) },
+                                    overlineContent = { NormalText(participant.bio) },
+                                    //supportingContent = {},
+                                    leadingContent = {
+                                        CustomProfilePictureFrame(
+                                            pictureUrl = participant.profilePicUrl,
+                                            frameColor = androidx.compose.ui.graphics.Color(
+                                                frameColors[participant.color]
+                                            )
+                                        )
+                                    },
+                                    //trailingContent = {},
+                                )
+                            }
+                        }
+                    }
                 }
             }
         }
@@ -152,7 +318,34 @@ fun GroupTrainingLobby(
 @Preview
 fun GroupTrainingLobbyPreview() {
     GroupTrainingLobby(
-        groupTrainingId = "JDq01mrnbav1pdaPuiy7",
-        onDeleteClick = {}
+        chosenGroupTraining = GroupTraining(
+            trainerId = "CnHNxTvP9UgT6Vbj8Xh8XFYcPmg1",
+            name = "Ranny beh",
+            trainingIndex = 0,
+            maxUsers = 4,
+            trainingDuration = 0,
+            timeDateOfTraining = 1732623634,
+            id = "04zGShmrMuJ0OekUfoR4",
+            numberOfParticipants = 1,
+            trainingDetails = "Toto su detaily treningu"
+        ),
+        onDeleteClick = {},
+        currentUser = User(
+            id = "",
+            displayName = "Pavol P",
+            profilePicUrl = "https://lh3.googleusercontent.com/a/ACg8ocK99KLif73J19brC6OLouuXT7I3ilQ-ArQApkXRHkuzZGOruVqf=s96-c",
+            birthDate = 0,
+            height = 0F,
+            weight = 0F,
+            bio = "Nie",
+            color = 0,
+            isTrainer = false
+        ),
+        onEndGroupTrainingClick = { },
+        allTrainingParticipants = emptyList(),
+        onCheckAllTrainingInfo = {
+            return@GroupTrainingLobby false
+        },
+        trainingState = {}
     )
 }
